requirements.txt
hidapi==0.7.99.post21
construct==2.9.45
Pillow==5.1.0
